# Optimize ST_AsMVTGeom for points