@javisantana 2017-07-25
PostGIS in general is full of things like this but nice catch