@bjornharrtell 2017-08-29
Thanks for this and https://carto.com/blog/inside/MVT-mapnik-vs-postgis. Very nice profiling guidance that I should remember.

With https://github.com/postgis/postgis/commit/47ac4b87b262cb5177c30e2834373777bbca69f1 I've made a more simple change that should result in more or less the same win as this PR. Further optimizations could be possible for the Point case as noted.

I've also made some other recent commits that should additionally reduce the amount of allocations and processing in general and optimized the case for LineString as I realized that only Polygons need validity/orientation correction. Polygon processing probably remains slow and would require non-trivial effort to optimize (I'm interested in the possibility of using https://github.com/mapbox/wagyu in PostGIS to replace the use of GEOS).